Change Snapshot to be an interface (tests aren't working yet)
