+++ pkg/analyze/snapshot/archivereader.go
@@ -1,20 +1,12 @@
 package snapshot
 
 import (
-	"context"
 	"fmt"
 	"io/fs"
-	storagev1 "k8s.io/api/storage/v1"
-	storagev1listers "k8s.io/client-go/listers/storage/v1"
-	"os"
 	"path/filepath"
 	"reflect"
 
-	scyllav1 "github.com/scylladb/scylla-operator/pkg/api/scylla/v1"
-	scyllav1listers "github.com/scylladb/scylla-operator/pkg/client/scylla/listers/scylla/v1"
-	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	corev1listers "k8s.io/client-go/listers/core/v1"
 	"k8s.io/client-go/tools/cache"
 )
 
@@ -28,8 +20,8 @@ func getIndexerForType(indexers map[reflect.Type]cache.Indexer, objType reflect.
 	return indexer
 }
 
-func NewSnapshotFromFS(fsys fs.FS, decoder runtime.Decoder) (*Snapshot, error) {
-	ds := NewSnapshot()
+func NewSnapshotFromFS(fsys fs.FS, decoder runtime.Decoder) (Snapshot, error) {
+	ds := NewDefaultSnapshot()
 
 	err := fs.WalkDir(fsys, ".", func(path string, d fs.DirEntry, err error) error {
 		if err != nil {
@@ -95,41 +87,4 @@ func IndexersFromFS(fsys fs.FS, decoder runtime.Decoder) (map[reflect.Type]cache
 		return nil, fmt.Errorf("can't walk the file tree: %w", err)
 	}
 	return indexers, nil
-}
-
-func NewDataSourceFromFS(ctx context.Context, archivePath string, decoder runtime.Decoder) (*DataSource, error) {
-	if decoder == nil {
-		return nil, fmt.Errorf("decoder must be specified")
-	}
-
-	indexers, err := IndexersFromFS(os.DirFS(archivePath), decoder)
-	if err != nil {
-		return nil, fmt.Errorf("can't build indexers from fs: %w", err)
-	}
-
-	types := []interface{}{
-		&corev1.Pod{},
-		&corev1.Service{},
-		&corev1.Secret{},
-		&corev1.ConfigMap{},
-		&corev1.ServiceAccount{},
-		&scyllav1.ScyllaCluster{},
-	}
-
-	// Create an indexer for each type if it wasn't created already
-	for _, obj := range types {
-		t := reflect.TypeOf(obj)
-		getIndexerForType(indexers, t)
-	}
-
-	return &DataSource{
-		PodLister:            corev1listers.NewPodLister(indexers[reflect.TypeOf(&corev1.Pod{})]),
-		ServiceLister:        corev1listers.NewServiceLister(indexers[reflect.TypeOf(&corev1.Service{})]),
-		SecretLister:         corev1listers.NewSecretLister(indexers[reflect.TypeOf(&corev1.Secret{})]),
-		ConfigMapLister:      corev1listers.NewConfigMapLister(indexers[reflect.TypeOf(&corev1.ConfigMap{})]),
-		ServiceAccountLister: corev1listers.NewServiceAccountLister(indexers[reflect.TypeOf(&corev1.ServiceAccount{})]),
-		ScyllaClusterLister:  scyllav1listers.NewScyllaClusterLister(indexers[reflect.TypeOf(&scyllav1.ScyllaCluster{})]),
-		StorageClassLister:   storagev1listers.NewStorageClassLister(getIndexerForType(indexers, reflect.TypeOf(&storagev1.StorageClass{}))),
-		CSIDriverLister:      storagev1listers.NewCSIDriverLister(getIndexerForType(indexers, reflect.TypeOf(&storagev1.CSIDriver{}))),
-	}, nil
 }
